refactor(wip): Change table scope location
