+++ modules/languages/python/src/main/java/org/vstu/meaningtree/languages/PythonLanguage.java
@@ -60,8 +60,8 @@ import org.vstu.meaningtree.nodes.types.containers.ListType;
 import org.vstu.meaningtree.nodes.types.containers.SetType;
 import org.vstu.meaningtree.nodes.types.containers.UnmodifiableListType;
 import org.vstu.meaningtree.nodes.types.user.Class;
+import org.vstu.meaningtree.utils.scopes.TypeScope;
 import org.vstu.meaningtree.utils.type_inference.HindleyMilner;
-import org.vstu.meaningtree.utils.type_inference.TypeScope;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayDeque;
